requirements.txt
appdirs==1.4.4
asgiref==3.3.4
autopep8==1.5.7
azure-common==1.1.27
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
black==21.6b0
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
Django==3.2.4
django-cors-headers==3.7.0
django-environ==0.4.5
django-extensions==3.1.3
django-storages==1.11.1
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.8.1
Pillow==8.2.0
pycodestyle==2.7.0
pycparser==2.20
PyJWT==2.1.0
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.5
uuid==1.30
